backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.0
python-dotenv==1.0.1

# Automation dependencies
playwright==1.48.0
openai==1.57.2
pillow==11.0.0
imagehash==4.3.1
beautifulsoup4==4.12.3
lxml==5.3.0

# WebSocket support
websockets==14.1

# Security
slowapi==0.1.9
cryptography==43.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
